+++ app/src/main/java/com/example/foodiedelivery/activities/LoginActivity.java
@@ -1,4 +1,4 @@
-package com.example.foodiedelivery.login;
+package com.example.foodiedelivery.activities;
 
 import androidx.activity.result.ActivityResult;
 import androidx.activity.result.ActivityResultCallback;
@@ -14,8 +14,10 @@ import android.util.Log;
 import android.view.View;
 import android.widget.ImageView;
 
-import com.example.foodiedelivery.activities.MainActivity;
 import com.example.foodiedelivery.R;
+import com.example.foodiedelivery.adapters.LoginAdapter;
+import com.example.foodiedelivery.models.User;
+import com.example.foodiedelivery.db.UserDbHelper;
 import com.google.android.gms.auth.api.signin.GoogleSignIn;
 import com.google.android.gms.auth.api.signin.GoogleSignInAccount;
 import com.google.android.gms.auth.api.signin.GoogleSignInClient;
