+++ app/src/main/java/com/example/foodiedelivery/adapters/ImageAdapter.java
@@ -1,4 +1,4 @@
-package com.example.foodiedelivery;
+package com.example.foodiedelivery.adapters;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -7,6 +7,8 @@ import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
 import androidx.viewpager2.widget.ViewPager2;
 
+import com.example.foodiedelivery.models.Image;
+import com.example.foodiedelivery.R;
 import com.makeramen.roundedimageview.RoundedImageView;
 
 import java.util.List;
