+++ app/src/main/java/com/example/foodiedelivery/adapters/RestaurantAdapter.java
@@ -1,4 +1,4 @@
-package com.example.foodiedelivery;
+package com.example.foodiedelivery.adapters;
 
 import android.view.Gravity;
 import android.view.LayoutInflater;
@@ -7,6 +7,9 @@ import android.view.ViewGroup;
 import android.widget.BaseAdapter;
 import android.widget.ImageView;
 import android.widget.TextView;
+
+import com.example.foodiedelivery.R;
+import com.example.foodiedelivery.models.Restaurant;
 
 import java.util.ArrayList;
 import java.util.List;
