+++ app/src/main/java/com/example/foodiedelivery/fragments/LoginFragment.java
@@ -1,4 +1,4 @@
-package com.example.foodiedelivery.login;
+package com.example.foodiedelivery.fragments;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -12,6 +12,8 @@ import androidx.fragment.app.Fragment;
 
 import com.example.foodiedelivery.activities.MainActivity;
 import com.example.foodiedelivery.R;
+import com.example.foodiedelivery.models.User;
+import com.example.foodiedelivery.db.UserDbHelper;
 
 public class LoginFragment extends Fragment {
