+++ app/src/main/java/com/example/foodiedelivery/fragments/LogoutFragment.java
@@ -1,4 +1,4 @@
-package com.example.foodiedelivery;
+package com.example.foodiedelivery.fragments;
 
 import android.os.Bundle;
 
@@ -7,6 +7,8 @@ import androidx.fragment.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
+import com.example.foodiedelivery.R;
 
 
 public class LogoutFragment extends Fragment {
