+++ app/src/main/java/com/example/foodiedelivery/models/Restaurant.java
@@ -1,4 +1,4 @@
-package com.example.foodiedelivery;
+package com.example.foodiedelivery.models;
 
 public class Restaurant {
     private String ResName;
