+++ app/src/main/res/layout/activity_image_upload.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ImageUploadActivity">
+    tools:context=".activities.ImageUploadActivity">
 
     <Button
         android:id="@+id/uploadImgBtn"
