+++ app/src/main/res/layout/activity_main.xml
@@ -8,7 +8,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:id="@+id/rel_layout"
-    tools:context=".MainActivity">
+    tools:context=".activities.MainActivity">
 
 
     <com.google.android.material.appbar.AppBarLayout
