+++ app/src/main/res/layout/fragment_home.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".HomeFragment">
+    tools:context=".fragments.HomeFragment">
 
     <androidx.viewpager2.widget.ViewPager2
         android:id="@+id/viewPager2id"
