+++ app/src/main/res/values/strings.xml
@@ -5,8 +5,8 @@
     <string name="signOutBtn">Sign Out</string>
     <!-- TODO: Remove or change this placeholder text -->
     <string name="hello_blank_fragment">Hello blank fragment</string>
-    <string name="txtRed">Red</string>
-    <string name="txtGreen">Green</string>
+    <string name="resBtn">Add Resturant</string>
+    <string name="menuBtn">Add Menu</string>
     <string name="txtLogout">LOG OUT</string>
     <string name="txtRestaurant">Restaurant</string>
     <string name="txtRedFragment">RED FRAGMENT HERE...</string>
